requirements.txt
setuptools>=38.5.1
typeguard>=2.7.0
humanfriendly
scipy>=1.4.1
pillow>=6.1.0
editdistance==0.5.2
ctc-segmentation>=1.4.0
wandb
# DNN related packages are installed by Makefile
# 'torch==1.0.1'
# "chainer==6.0.0",
# 'cupy==6.0.0',
tensorboard>=1.14
# Signal processing related
librosa>=0.8.0
# Natural language processing related
# FIXME(kamo): Sentencepiece 0.1.90 breaks backwardcompatibility?
sentencepiece<0.1.90,>=0.1.82
nltk>=3.4.5
# File IO related
PyYAML>=5.1.2
soundfile>=0.10.2
h5py>=2.10.0
kaldiio>=2.17.0
# TTS related
pyworld>=0.2.10
espnet_tts_frontend
# ASR frontend related
nara_wpe>=0.0.5
torch_complex
pytorch_wpe
matplotlib==3.3.3
parallel_wavegan==0.4.8
flask